pyproject.toml
[project]
name = "tutoriais"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.3",
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "gunicorn>=23.0.0",
    "markdown>=3.8",
    "pillow>=11.2.1",
    "python-decouple>=3.8",
    "whitenoise[brotli]>=6.9.0",
]

[dependency-groups]
dev = [
    "django-extensions>=4.1",
    "ipython>=9.3.0",
    "model-bakery>=1.20.5",
    "pytest>=8.4.0",
    "pytest-django>=4.11.1",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "backend.config.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
